[Refac] Creating city_name_helpers.py
