+++ covid19br/common/models/bulletin_models.py
@@ -1,12 +1,7 @@
 import datetime
 from abc import ABC
-from functools import lru_cache
-from pathlib import Path
 
-import rows
-from rows.fields import slug
-from fuzzy_types import FuzzyDict
-
+from covid19br.common.city_name_helpers import fix_city_name
 from covid19br.common.constants import (
     IMPORTED_OR_UNDEFINED_CODE,
     NOT_INFORMED_CODE,
@@ -15,65 +10,6 @@ from covid19br.common.constants import (
 )
 from covid19br.common.data_normalization_utils import NormalizationUtils
 from covid19br.common.exceptions import BadReportError
-
-
-BASE_PATH = Path(__file__).parent.parent.parent
-POPULATION_PATH = BASE_PATH / "data" / "populacao-por-municipio-2020.csv"
-
-@lru_cache(maxsize=6000)
-def normalize_city_name(city):
-    city = slug(city)
-
-    for word in ("da", "das", "de", "do", "dos"):
-        city = city.replace(f"_{word}_", "_")
-
-    return city
-
-
-@lru_cache()
-def brazilian_population():
-    return rows.import_from_csv(POPULATION_PATH)
-
-
-@lru_cache(maxsize=27)
-def state_population(state):
-    return [row for row in brazilian_population() if row.state == state]
-
-
-@lru_cache(maxsize=900)
-def city_id_from_name(state):
-    data = {row.city: int(row.city_ibge_code) for row in state_population(state)}
-    data["Importados/Indefinidos"] = None
-    return data
-
-
-@lru_cache(maxsize=900)
-def get_city_id_from_name(state, name):
-    normalized_cities = FuzzyDict({
-        normalize_city_name(key): value
-        for key, value in city_id_from_name(state).items()
-    })
-    return normalized_cities[normalize_city_name(name)]
-
-
-@lru_cache(maxsize=27)
-def city_name_from_id(state):
-    return {value: key for key, value in city_id_from_name(state).items()}
-
-
-@lru_cache(maxsize=5570)
-def get_city_name_from_id(state, city_id):
-    return city_name_from_id(state)[city_id]
-
-
-def fix_city_name(state, city):
-    try:
-        city_id = get_city_id_from_name(state, city)
-        city_name = get_city_name_from_id(state, city_id)
-    except KeyError:
-        raise ValueError(f"Unknown city '{city}' for state {state}")
-    else:
-        return city_name
 
 
 class BulletinModel(ABC):
